Add files via upload
Online retail is a transnational data set which contains all the transactions occurring between 01/12/2010 and 09/12/2011 for a UK-based and registered non-store online retail. The company mainly sells unique all-occasion gifts. Many customers of the company are wholesalers.

The steps are broadly:

Read and understand the data
Clean the data
Prepare the data for modelling
Modelling
Final analysis and recommendations.